connecting the front end and backend. Fixed the root warning
